+++ header.php
@@ -31,13 +31,13 @@
       <div class="site-header__wrapper font-color-white font-size-10 font-size-s-8 font-weight-bold">
         <div class="site-header__main container">
           <div class="grid-row">
-            <nav class="grid-item is-xxl-5 layout-flex-align-center" role="navigation" aria-label="Main" data-testid="site-nav">
+            <nav class="grid-item is-s-6 is-xxl-5 layout-flex-align-center" role="navigation" aria-label="Main" data-testid="site-nav">
               <ul class="site-header__navigation u-inline-list u-inline-block">
                 <li class="site-header__toggle site-header__nav-toggle ux-pointer" role="button" tabindex="0" aria-controls="header-sub" aria-label="Site Navigation" aria-haspopup="menu" aria-pressed="false"><i class="icon-menu icon-large"></i></li>
                 <li class="site-header__toggle site-header__search-toggle ux-pointer" role="button" tabindex="0" aria-controls="header-search" aria-label="Search" aria-haspopup="dialog" aria-pressed="false"><i class="icon-search icon-large"></i></li>
               </ul>
             </nav>
-            <div class="header-main__middle grid-item is-xxl-14 text-align-center">
+            <div class="header-main__middle grid-item is-s-12 is-xxl-14 text-align-center">
               <a href="<?php echo home_url(); ?>">
                 <nav class="site-header__logomark u-inline-block"><?php echo nm_get_file( '/dist/img/logomark-white.svg' ); ?></nav>
                 <div class="site-header__scroll-reveal font-size-9 font-size-s-8">